requirements.txt
apt-clone==0.2.1
apturl==0.5.2
attrs==21.2.0
Babel==2.8.0
beautifulsoup4==4.10.0
blinker==1.7.0
Brlapi==0.8.3
cachetools==5.3.2
catfish==4.16.3
ccsm==0.9.14.1
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
compizconfig-python==0.9.14.1
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
curl_cffi==0.6.0b9
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duckduckgo_search==4.4.3
eyeD3==0.8.10
filelock==3.6.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
google-api-core==2.17.1
google-api-python-client==2.48.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
Google-Images-Search==1.4.6
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.30.2
httplib2==0.20.2
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==5.1.0
Mako==1.1.3
MarkupSafe==2.1.5
menulibre==2.2.2
more-itertools==8.10.0
mugshot==0.4.3
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
OdooRPC==0.10.1
onboard==1.4.1
packaging==21.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
pyfiglet==0.8.post1
Pygments==2.11.2
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyparted==3.11.7
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu3
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-magic==0.4.24
python-resize-image==1.1.20
python-xapp==2.2.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
rsa==4.9
SecretStorage==3.3.1
setproctitle==1.2.2
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==2.0.27
systemd-python==234
termcolor==1.1.0
texttable==1.6.4
tldextract==3.1.2
typing_extensions==4.9.0
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
uritemplate==4.1.1
urllib3==1.26.5
wadllib==1.3.6
websocket-client==1.2.3
Werkzeug==3.0.1
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
zipp==1.0.0
